restructure ui - forms
