+++ apps/app/src/app/(protected)/events/create/form/event-roles/index.tsx
@@ -16,7 +16,7 @@ import { IcreateEventSchema } from '../validation';
 import { IeventRoleSchema } from './validation';
 import { IeventDetailsSchema } from '../event-info/validation';
 import { LoaderSpinner } from '__components/ui/LoaderSpinner';
-import { getAttributes } from 'ui/src/forms/attributesMap';
+import { getAttributes } from 'ui/src/form-elements/attributesMap';
 import { BiEdit, BiTrash } from 'react-icons/bi';
 
 type Props = {
